refactor: 채널 리스트 컴포넌트 분리 #128
